+++ example/lib/main.dart
@@ -1,7 +1,0 @@
-import 'package:example/example_flutter_app.dart';
-import 'package:flutter/material.dart';
-
-void main() {
-  enableLogging();
-  runApp(const MaterialApp(home: OdbcExampleApp()));
-}